# fix: check proof state before fetch