@jleach 2024-11-12
> Reason for getting rid of the `Promise.all` and making them run consecutively?

No good reason. I though it looked cleaner. I gave it a test and it does run slightly faster with it so I added it back.

Without Promise All: 489 ms, 393 ms
With Promise All: 307 ms, 325 ms